requirements.txt
fastapi[all]
sqlalchemy
alembic
PyMySQL[rsa]
Pillow
numpy